requirements.txt
asgiref==3.8.1
astroid==3.3.6
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
dill==0.3.9
django-environ==0.11.2
django-phonenumber-field==8.0.0
Django==4.2
django-ratelimit==4.1.0
django-recaptcha==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.52
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.2
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pyOpenSSL==24.3.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.2
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
zope.interface==7.2
channels_redis==4.2.1
selenium==4.27.1
django-cryptography==1.1
factory_boy==3.3.1
drf-yasg==1.21.8
pillow==11.0.0
django_phonenumber_field==8.0.0
phonenumbers==8.13.52
djoser==2.3.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-filter==24.3
mongoengine==0.29.1
django-rest-framework-mongoengine==3.4.1


